# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DocTrackr API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /document-types:
    get:
      operationId: listDocumentTypes
      tags: [document-types]
      summary: List document types
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [student, employee, both, all]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, all]
      responses:
        "200":
          description: List of document types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentType"
    post:
      operationId: createDocumentType
      tags: [document-types]
      summary: Create document type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentTypeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentType"
  /document-types/{id}:
    get:
      operationId: getDocumentType
      tags: [document-types]
      summary: Get document type by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentType"
    put:
      operationId: updateDocumentType
      tags: [document-types]
      summary: Update document type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentTypeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentType"
    delete:
      operationId: deleteDocumentType
      tags: [document-types]
      summary: Delete document type (only if unused)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted or blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /document-types/{id}/toggle-active:
    patch:
      operationId: toggleDocumentTypeActive
      tags: [document-types]
      summary: Activate or deactivate a document type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentType"

  # Checklist Items
...
  /dashboard/expiring-soon:
    get:
      operationId: getDashboardExpiringSoon
      tags: [dashboard]
      summary: Get items expiring in next 30 days
      parameters:
        - name: personType
          in: query
          schema:
            type: string
            enum: [student, employee, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expiring items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistItem"
  /dashboard/overdue:
    get:
      operationId: getDashboardOverdue
      tags: [dashboard]
      summary: Get overdue items
      parameters:
        - name: personType
          in: query
          schema:
            type: string
            enum: [student, employee, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Overdue items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistItem"
  /dashboard/missing:
    get:
      operationId: getDashboardMissing
      tags: [dashboard]
      summary: Get missing document items
      parameters:
        - name: personType
          in: query
          schema:
            type: string
            enum: [student, employee, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Missing document items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistItem"
components:
  schemas:
...
    ActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    DocumentType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [student, employee, both]
        description:
          type: ["string", "null"]
        isRequired:
          type: boolean
        isActive:
          type: boolean
        templateFormUrl:
          type: ["string", "null"]
        usageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - isRequired
        - isActive
        - usageCount
        - createdAt
        - updatedAt
    DocumentTypeInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [student, employee, both]
        description:
          type: ["string", "null"]
        isRequired:
          type: boolean
        isActive:
          type: boolean
      required:
        - name
        - category
        - isRequired
    ChecklistItem:
      type: object
      properties:
        id:
          type: integer
        personType:
          type: string
          enum: [student, employee]
        personId:
          type: integer
        personName:
          type: string
        personEmail:
          type: ["string", "null"]
        personStatus:
          type: string
          enum: [active, inactive]
        documentTypeId:
          type: integer
        documentTypeName:
          type: string
        documentTypeCategory:
          type: string
        templateFormUrl:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [missing, valid, expiring_soon, due_today, overdue]
        notes:
          type: ["string", "null"]
        lastReminderSentAt:
          type: ["string", "null"]
        nextReminderDueAt:
          type: ["string", "null"]
        daysUntilExpiry:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - personType
        - personId
        - personName
        - personStatus
        - documentTypeId
        - documentTypeName
        - documentTypeCategory
        - status
        - createdAt
        - updatedAt
...